+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add tests for FullTextRequestUsageTable component. (UIEH-1179)
 * Add tests for SummaryTable component. (UIEH-1180)
 * Add tests for ApplicationRoute component. (UIEH-1184)
+* Add tests for TitlesTable component. (UIEH-1181)
 
 ## [7.0.1] (https://github.com/folio-org/ui-eholdings/tree/v7.0.1) (2021-10-19)